Updates alternative install directions
I noticed that the alternative install instructions which update an existing Conda environment did not work because it would reference the 'name' parameter defined in the environment.yaml file. Added a '--name' flag to the command to tell Conda to ignore that parameter and instead update the User's existing environment. Tested on my local machine.